Warnings solved
git-svn-id: https://svn.code.sf.net/p/tcnopen/trdp/trunk@1021 3b5a3598-5f4e-4449-9e63-bd40438bfec0

+++ trdp/src/vos/windows/vos_shared_mem.c
@@ -195,7 +195,7 @@ EXT_DECL VOS_ERR_T vos_sharedOpen (
     vos_memFree(shMemName);
     if (retVal != VOS_NO_ERR)
     {
-        *pSize = NULL;
+        *pSize = (UINT32)NULL;
     }
     else
     {
